Fixed AckOverflow handling. Fixed runaway in reliable socket when going over 255 transactions.

+++ Acn/Citp/Packets/CaEx/CaExLiveViewStatus.cs
@@ -53,12 +53,12 @@ namespace Citp.Packets.CaEx
             Availability = (LiveViewAvailability)data.ReadByte();
             Width = data.ReadUInt16();
             Height = data.ReadUInt16();
-            CameraPositionX = data.ReadByte();
-            CameraPositionY = data.ReadByte();
-            CameraPositionZ = data.ReadByte();
-            CameraFocusX = data.ReadByte();
-            CameraFocusY = data.ReadByte();
-            CameraFocusZ = data.ReadByte();
+            CameraPositionX = data.ReadSingle();
+            CameraPositionY = data.ReadSingle();
+            CameraPositionZ = data.ReadSingle();
+            CameraFocusX = data.ReadSingle();
+            CameraFocusY = data.ReadSingle();
+            CameraFocusZ = data.ReadSingle();
         }
 
         public override void WriteData(CitpBinaryWriter data)
